requirements.txt
Fabric==1.3.8
Flask==0.10.1
Frozen-Flask==0.11
Jinja2==2.7.2
Markdown==2.4
MarkupSafe==0.19
PyYAML==3.10
Pygments==1.5
Werkzeug==0.9.4
argparse==1.2.1
docutils==0.10
ipdb==0.8
ipython==2.0.0
itsdangerous==0.24
python-dateutil==1.5
wsgiref==0.1.2
